requirements.txt
future==0.15.2
Jinja2==2.8
klein==15.3.1
MarkupSafe==0.23
pyotp==2.1.1
Twisted==16.3.0
Werkzeug==0.11.10
zope.interface==4.2.0
